requirements.txt
discord.py==2.3.2
requests==2.32.3
python-dotenv==1.0.0
web3==6.15.1 